APP: Add console redirection when running applications
This adds redirection of the output (terminal) of the app that is
run with ?run query param. A special helper script will follow to
help facilitate a working solution.

Currently it hardcodes UDP ports 33332/33334 for its use.

Redirection uses UDP because it's state-less nature means lower overhead
for the network handling.

Loss of packets in not an issue on a local network.

Redirection is done by interposing BIOS console routines
AND GEMDOS Cconws(). GEMDOS interposition is only added for
performance reasonds because Cconws allows for entire strings
to be sent in a single packet, which means lower overhead.

Overall redirection should be lower overhead then printing to
the console because console scrolling means copying video memory.