SCRIPTS: Add uip-run.py script
This script facilitates running a TOS or TTP app and redirects
console while doing it.

This is meant as a development aid for coders and I mainly thought
about myself while working on it.

It's not complete yet. Loads of scancodes are not translated and
some escape codes are not translated either.