Fix calling convention issue that could prevent device detection
Note the different calling conventions: USB stack is compiled with -mshort,
whereas uIP needs to be built without -mshort.

Hence, the 'ifnum' parameter is not passed correctly, when the USB stack
calls ethernet_probe(). Luckily, we know that both Asix and PicoWifi
only have interface 0, and can therefore hardcode it.

Thank you to Claude Labelle for reporting this issue.